refactor: Move test helper functions to utils file
